Added a DAO Models sectoin
Just listened to the zeroknowledge interview with @owocki. Was really inspired, but disappointed that DAOs weren't part of that conversation as DAOs are a great tool to help communities/projects make decisions and manage money. Created a section to talk about that. I added it in after the Continuous Funding section because ideally DAOs would be self sufficient, community driven, and around for a long time since the DAO itself owns it's assets rather than a CEO or project founder.